chore: typings
