+++ packages/core/src/index.ts
@@ -12,5 +12,5 @@ export { sharedHooks } from './shared-hooks'
 export { getStaticCss } from './static-css'
 export { Stylesheet, type StylesheetOptions } from './stylesheet'
 export { toCss } from './to-css'
-export type { StylesheetContext } from './types'
+export type { RecipeNode, StylesheetContext, TransformResult } from './types'
 export { Utility } from './utility'
