+++ packages/generator/src/engines/jsx.ts
@@ -20,5 +20,5 @@ export interface PandaJsxEngine {
   typeName: string
   componentName: string
   framework: UserConfig['jsxFramework']
-  styleProps: 'all' | 'minimal' | 'none'
+  styleProps: UserConfig['jsxStyleProps']
 }
